move TractActivity intent creation/launch logic to :ui:base module
