+++ app/src/main/java/org/cru/godtools/util/ActivityUtils.kt
@@ -3,9 +3,9 @@ package org.cru.godtools.util
 import android.app.Activity
 import java.util.Locale
 import org.cru.godtools.base.tool.startLessonActivity
-import org.cru.godtools.base.tool.startTractActivity
 import org.cru.godtools.base.ui.startArticlesActivity
 import org.cru.godtools.base.ui.startCyoaActivity
+import org.cru.godtools.base.ui.startTractActivity
 import org.cru.godtools.model.Tool.Type
 
 fun Activity.openToolActivity(code: String, type: Type, vararg languages: Locale, showTips: Boolean = false) =
