+++ ui/base-tool/src/main/kotlin/org/cru/godtools/base/tool/Activities.kt
@@ -6,11 +6,8 @@ import android.content.Intent
 import android.os.Bundle
 import java.util.Locale
 import org.ccci.gto.android.common.util.os.putLocale
-import org.ccci.gto.android.common.util.os.putLocaleArray
 import org.cru.godtools.base.EXTRA_LANGUAGE
-import org.cru.godtools.base.EXTRA_LANGUAGES
 import org.cru.godtools.base.EXTRA_TOOL
-import org.cru.godtools.base.ui.EXTRA_SHOW_TIPS
 
 const val SHORTCUT_LAUNCH = "org.cru.godtools.tool.SHORTCUT_LAUNCH"
 
@@ -25,21 +22,3 @@ fun Context.createLessonActivityIntent(toolCode: String, language: Locale) =
         .putExtra(EXTRA_TOOL, toolCode)
         .putExtras(Bundle().apply { putLocale(EXTRA_LANGUAGE, language) })
 // endregion LessonActivity
-
-// region TractActivity
-private const val ACTIVITY_CLASS_TRACT = "org.cru.godtools.tract.activity.TractActivity"
-
-fun Activity.startTractActivity(toolCode: String, vararg languages: Locale?, showTips: Boolean) =
-    startActivity(createTractActivityIntent(toolCode, *languages, showTips = showTips))
-
-fun Context.createTractActivityIntent(toolCode: String, vararg languages: Locale?, showTips: Boolean = false) =
-    Intent().setClassName(this, ACTIVITY_CLASS_TRACT)
-        .putExtra(EXTRA_TOOL, toolCode)
-        .putLanguagesExtra(*languages)
-        .putExtra(EXTRA_SHOW_TIPS, showTips)
-
-private fun Intent.putLanguagesExtra(vararg languages: Locale?) = putExtras(
-    // XXX: we use singleString mode to support using this intent for legacy shortcuts
-    Bundle().apply { putLocaleArray(EXTRA_LANGUAGES, languages.filterNotNull().toTypedArray(), true) }
-)
-// endregion TractActivity
